refactor devise related auth behaviors into separated concern
