+++ app/models/user.rb
@@ -2,30 +2,18 @@ class User
   INTERACTIVE_IDENTITY_VERIFICATION_SUCCESS_CODE = "acc"
   MIN_USERNAME_LENGTH = 8
   MAX_USERNAME_LENGTH = 60
-  MAX_SAME_CHAR_LIMIT = 4
   include Mongoid::Document
   include Mongoid::Timestamps
   include Acapi::Notifiers
+  include AuthorizationConcerns
 
   attr_accessor :login
 
-  # Include default devise modules. Others available are:
-  # :confirmable, :lockable, :timeoutable and :omniauthable
-  devise :database_authenticatable, :registerable, :lockable,
-         :recoverable, :rememberable, :trackable, :timeoutable, :authentication_keys => {email: false, login: true}
-
   validates_presence_of :oim_id
   validates_uniqueness_of :oim_id, :case_sensitive => false
-  validate :password_complexity
   validate :oim_id_rules
   validates_uniqueness_of :email,:case_sensitive => false
-  validates_presence_of     :password, if: :password_required?
-  validates_confirmation_of :password, if: :password_required?
-  validates_length_of       :password, within: Devise.password_length, allow_blank: true
-  validates_format_of :email, with: Devise::email_regexp , allow_blank: true, :message => "(optional) is invalid"
 
-  scope :locked, ->{ where(:locked_at.ne => nil) }
-  scope :unlocked, ->{ where(locked_at: nil) }
   scope :datatable_search, ->(query) { self.where({"$or" => ([{"name" => Regexp.compile(Regexp.escape(query), true)}, {"email" => Regexp.compile(Regexp.escape(query), true)}])}) }
 
   def oim_id_rules
@@ -38,57 +26,9 @@ class User
     end
   end
 
-  def password_complexity
-    if password.present? and not password.match(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^a-zA-Z\d ]).+$/)
-      errors.add :password, "must include at least one lowercase letter, one uppercase letter, one digit, and one character that is not a digit or letter or space"
-    elsif password.present? and password.match(/#{Regexp.escape(oim_id)}/i)
-      errors.add :password, "cannot contain username"
-    elsif password.present? and password_repeated_chars_limit(password)
-      errors.add :password, "cannot repeat any character more than #{MAX_SAME_CHAR_LIMIT} times"
-    elsif password.present? and password.match(/(.)\1\1/)
-      errors.add :password, "must not repeat consecutive characters more than once"
-    elsif password.present? and !password.match(/(.*?[a-zA-Z]){4,}/)
-      errors.add :password, "must have at least 4 alphabetical characters"
-    end
-  end
-
-  def password_repeated_chars_limit(password)
-    return true if password.chars.group_by(&:chr).map{ |k,v| v.size}.max > MAX_SAME_CHAR_LIMIT
-    false
-  end
-
-  def password_required?
-    !persisted? || !password.nil? || !password_confirmation.nil?
-  end
-
   def valid_attribute?(attribute_name)
     self.valid?
     self.errors[attribute_name].blank?
-  end
-
-  def self.password_invalid?(password)
-    user = User.new(oim_id: 'example1', password: password)
-    !user.valid_attribute?('password')
-  end
-
-  def self.generate_valid_password
-    password = Devise.friendly_token.first(16)
-    password = password + "aA1!"
-    password = password.squeeze
-    if password_invalid?(password)
-      password = generate_valid_password
-    else
-      password
-    end
-  end
-
-  def self.find_for_database_authentication(warden_conditions)
-    conditions = warden_conditions.dup
-    if login = conditions.delete(:login).downcase
-      where(conditions).where('$or' => [ {:oim_id => /^#{Regexp.escape(login)}$/i}, {:email => /^#{Regexp.escape(login)}$/i} ]).first
-    else
-      where(conditions).first
-    end
   end
 
   def switch_to_idp!
@@ -116,14 +56,6 @@ class User
   ## Seed: https://github.com/plataformatec/devise/wiki/How-To%3a-Require-admin-to-activate-account-before-sign_in
   field :approved, type: Boolean, default: true
 
-  ## Database authenticatable
-  field :email,              type: String, default: ""
-  field :encrypted_password, type: String, default: ""
-
-  ## Recoverable
-  field :reset_password_token,   type: String
-  field :reset_password_sent_at, type: Time
-
   ##RIDP
   field :identity_verified_date, type: Date
   field :identity_final_decision_code, type: String
@@ -131,18 +63,9 @@ class User
   field :identity_response_code, type: String
   field :identity_response_description_text, type: String
 
-  ## Rememberable
-  field :remember_created_at, type: Time
-
   ## Trackable
-  field :sign_in_count,      type: Integer, default: 0
-  field :current_sign_in_at, type: Time
-  field :last_sign_in_at,    type: Time
-  field :current_sign_in_ip, type: String
-  field :last_sign_in_ip,    type: String
   field :idp_uuid, type: String
 
-  field :authentication_token
   field :roles, :type => Array, :default => []
 
   # Oracle Identity Manager ID
@@ -150,11 +73,6 @@ class User
 
   field :last_portal_visited, type: String
   field :idp_verified, type: Boolean, default: false
-
-  ## Lockable
-  field :failed_attempts, type: Integer, default: 0 # Only if lock strategy is :failed_attempts
-  field :unlock_token,    type: String # Only if unlock strategy is :email or :both
-  field :locked_at,       type: Time
 
   index({preferred_language: 1})
   index({approved: 1})
@@ -188,20 +106,6 @@ class User
   accepts_nested_attributes_for :person, :allow_destroy => true
 
   # after_initialize :instantiate_person
-
-  ## Confirmable
-  # field :confirmation_token,   type: String
-  # field :confirmed_at,         type: Time
-  # field :confirmation_sent_at, type: Time
-  # field :unconfirmed_email,    type: String # Only if using reconfirmable
-
-  ## Lockable
-  field :failed_attempts, type: Integer, default: 0 # Only if lock strategy is :failed_attempts
-  field :unlock_token,    type: String # Only if unlock strategy is :email or :both
-  field :locked_at,       type: Time
-
-  before_save :ensure_authentication_token
-
   #  after_create :send_welcome_email
 
   delegate :primary_family, to: :person, allow_nil: true
@@ -237,24 +141,6 @@ class User
   def send_welcome_email
     UserMailer.welcome(self).deliver_now
     true
-  end
-
-  def set_random_password(passwd)
-    self.password = passwd
-    self.password_confirmation = passwd
-    self.save
-  end
-
-  def update_lockable
-    if locked_at.nil?
-      self.lock_access!
-    else
-      self.unlock_access!
-    end
-  end
-
-  def lockable_notice
-    self.locked_at.nil? ? 'unlocked' : 'locked'
   end
 
   def has_role?(role_sym)
@@ -342,13 +228,6 @@ class User
     person == employer_profile.active_broker if employer_profile.active_broker
   end
 
-  def ensure_authentication_token
-    if authentication_token.blank?
-      self.authentication_token = generate_authentication_token
-    end
-    true
-  end
-
   def instantiate_person
     self.person = Person.new
   end
@@ -359,20 +238,6 @@ class User
   def self.orphans
     user_ids = Person.where(:user_id => { "$ne" => nil }).pluck(:user_id)
     User.where("_id" => { "$nin" => user_ids }).order(email: :asc).entries
-  end
-
-  def self.send_reset_password_instructions(attributes={})
-    recoverable = find_or_initialize_with_errors(reset_password_keys, attributes, :not_found)
-    if !recoverable.approved?
-      recoverable.errors[:base] << I18n.t("devise.failure.not_approved")
-    elsif recoverable.persisted?
-      recoverable.send_reset_password_instructions
-    end
-    recoverable
-  end
-
-  def self.find_by_authentication_token(token)
-    where(authentication_token: token).first
   end
 
   def handle_headless_records
@@ -505,12 +370,5 @@ class User
   def strip_empty_fields
     unset("email") if email.blank?
     unset("oim_id") if oim_id.blank?
-  end
-
-  def generate_authentication_token
-    loop do
-      token = Devise.friendly_token
-      break token unless User.where(authentication_token: token).first
-    end
   end
 end
